+++ Changelog.md
@@ -23,6 +23,7 @@
 - Fix: The Keypad Door can only be accessed with an empty hand
 - Fix: Several blocks do not respect team ownership
 - Fix: Server crash involving automatically building a Block Pocket
+- Fix: Tricking scanners with player heads does not work if the player head owner is on the allowlist of the block
 - Removed: Some unnecessary menu types. This may result in a "missing registry entries" message showing up, which can be accepted
 - Removed: "allowCodebreakerItem" config option. Disabling the Codebreaker is now achieveable by setting "codebreaker_chance" to a negative value
 - Misc.: More texture and model changes